docs(DevTools): Add info about binding state manipulation
This is related to https://github.com/aurelia/binding/pull/714, and you
can include my change in your docs, include a link to the aurelia-binding
document or do nothing if you think it does not warrant a document entry.
However, i did want to point this out because it tripped me up when I was
comparing my initial state to my current state. Redux DevTools was showing my
state as equal and my logic was alerting me that my state was not equal because
of aurelia-binding observing properties (i was counting properties to check
for equality before checking every value).

I found this using redux and not this library, but I do plan on making
the switch as soon as I get time. To make sure it behaved the same
in this library, i replicated it in the contacts example by adding a
`foobar` input in the `contact-detail.html`