feat(devtools): Add action params to the devtools Action view
Originally, only the `type` property was sent to the dev tools. This
commit adds the `params` to the private action object so the type and
the params are shown on the Action view in the devtools. IMPORTANT: anything that is
not serializable (aka a function) and is dispatched as a param will
show as a `null` value.

closes #75