feat(rxjs) update `rxjs` to be a peerDependency (v6 or v7)
Allows consuming projects to upgrade to rxjs 7 without a version
resolution override for aurelia-store.
Without this change, build systems would bundle both versions
 and warn of the duplication.
Modifications:
- Install RXJS 7 as a devDependency for the local build
- Update TypeScript to 4.3 for compatibility with RXJS 7
- Update other build devDependencies to fix build errors with new versions

Possible breaking changes:
`rxjs` would have to be installed in the consuming project, which might not be
automatic with older versions of npm. Anyone using the plugin would
likely already be using RXJS to interact with the store, though.